requirements.txt
accelerate
brouhaha @ https://github.com/marianne-m/brouhaha-vad/archive/main.zip#sha256=3e920f66726a95dd09249a056051732dce44d08e3c2005f20aa9e9b84455376d
datasets<3.0.0,>=2.16.1
deepgram_captions
deepgram_sdk
demucs
g2p
huggingface_hub
llm-swarm
matplotlib
multiprocess
noisereduce
numpy
pandas
penn
pyannote.audio
pyannote.core
pyannote.database
pyannote.metrics
pyannote.pipeline
pyarrow
pydub
pyloudnorm
python-dotenv
setuptools
soundfile
torch
torchaudio
tqdm
transformers
